IsInside för polygon

+++ Projektarbete_grupp 17/Koordinat.cs
@@ -8,11 +8,11 @@ namespace Projektarbete_grupp_17
 {
     class Koordinat
     {
-        public int y;
-        public int x;
+        public double y;
+        public double x;
        
 
-        public Koordinat (int inputx, int inputy)
+        public Koordinat (double inputx, double inputy)
         {
             y = inputy;
             x = inputx;
